[+] Small fixes

+++ dlpf/gym_environ/policies/path.py
@@ -62,7 +62,7 @@ class RandomStartAndFinishMixin(object):
                 if local_map[self.start] == 0 \
                         and local_map[self.finish] == 0 \
                         and self.start != self.finish \
-                        and check_finish_achievable(numpy.array(local_map, dtype=numpy.int),
+                        and check_finish_achievable(numpy.array(local_map, dtype=numpy.float),
                                                     numpy.array(self.start, dtype=numpy.int),
                                                     numpy.array(self.finish, dtype=numpy.int)):
                     break
